very nice linker relaxation thank you riscv
I got it! (the assembler address thing)

So the problem is actually not with the assembler.
I `objdump` upon the dot o files from the output of `as` and that actually got the address correctly. It's just the binary (or when i check in gdb it won't work.

If you remember one of the comic on Chris Dutchyn's wall where it says something like "the problem happens not AFTER the writing, but DURING the writing"), it really happened.

I can tell you the full story tomorrow, but the short version is that it I figured it out when I was half-way writing the stack-overflow question.

https://stackoverflow.com/questions/77797870/correct-way-to-add-labels-for-riscv-data-section-so-assembler-can-pick-up

So `ld` has something called "relaxation rules", which are like "global scope optimization happens in linking stage". and are machine-dependent, and are enabled by default if the machine supports it.

RISCV has one of the relaxations using global pointers

https://www.sifive.com/blog/all-aboard-part-3-linker-relaxation-in-riscv-toolchain

short version:

https://maskray.me/blog/2021-03-14-the-dark-side-of-riscv-linker-relaxation#global-pointer-relaxation

So it goes like this:

usually when we need to get an address to a register, since instruction size is limiting the size of immediates of I-type instruction, it need 2 instructions (and that's why `la` is the pseudo instruction), first to get the upper-half of the address to the register (usually with `auipc`), then getting the lower half by `addi` the offset.

but since most things in data sections are relatively close, so when accessing them, the `auipc` part are identical and only the offset differs.

So it turns out that you'd only need 1 `auipc`, and get other addresses just by adding offset (hopefully it's within the range of how much immediate can take) to the base address.

That's why the register `gp` (first when i see it uses gp for address i thought it's just as a placeholder where the assembler can't "see" as it lacks some additional labels, but now it doesn't seems to be the case).

In the crt0.o, which i read the code in the summer and forgot about it only remembers that "oh so that's how it got `main` as entry point" part.

So `gp` was set in crt0.o to the BSS section, and as a result, every access to the `.data` section in the future, does not require the `auipc` part but just `addi` relative to `gp`, given if `gp` has been set during the linking stage.

Which, as we are using `gcc` with `--nostartfiles` or `ld` without specify crt0.o and libc (it needs memset in crt0.o to clear the memory for BSS). That relaxation is now a bad idea since gp is not defined, so every access relative to gp will cause a segfault.

tl;dr

giving `--no-relax` flag to `ld` should fix the problem.

I tested, it worked now.